+++ linetools/guis/tests/test_guis.py
@@ -10,7 +10,7 @@ import pytest
 import numpy as np
 from astropy import units as u
 
-from PyQt5.QtWidgets import QApplication
+from qtpy.QtWidgets import QApplication
 
 from linetools.guis import xspecgui, xabssysgui
 from linetools.guis import utils as ltgu