+++ linetools/spectra/xspectrum1d.py
@@ -671,7 +671,7 @@ class XSpectrum1D(object):
         # Launch XSpectrum1D??
         if 'xspec' in kwargs:
             import sys
-            from PyQt5.QtWidgets import QApplication
+            from qtpy.QtWidgets import QApplication
             from linetools.guis.xspecgui import XSpecGui
             app = QApplication(sys.argv)
             # Scale to pixels on screen
@@ -1366,7 +1366,7 @@ class XSpectrum1D(object):
         print('Wrote spectrum to {:s}'.format(outfil))
 
     def fit_continuum(self, knots=None, edges=None, wlim=None, dw=10.,
-                      kind=None, numguesspix=10, **kwargs):
+                      kind=None, numguesspix=10, outfknots=None, **kwargs):
         """ Interactively fit a continuum.
 
         This sets the following attributes
@@ -1398,6 +1398,8 @@ class XSpectrum1D(object):
         numguesspix : int, optional
           Number of pixels included when guessing knot location using flux
           median ('A' or 'M'); default is 10
+        outfknots : str
+          Output json file where knots will be written.
         **kwargs : dict
           Other keyword arguments are passed to
           ~linetools.analysis.continuum.find_continuum.  For
@@ -1433,7 +1435,7 @@ class XSpectrum1D(object):
         else:
             if edges is None:
                 nchunks = max(3, (wmax - wmin) / float(dw))
-                edges = np.linspace(wmin, wmax, nchunks + 1)
+                edges = np.linspace(wmin, wmax, np.int(nchunks) + 1)
 
         if knots is None:
             knots, indices, masked = prepare_knots(
@@ -1460,7 +1462,7 @@ class XSpectrum1D(object):
         fig = plt.figure(figsize=(11, 7))
         fig.subplots_adjust(left=0.05, right=0.95, bottom=0.1, top=0.95)
         wrapper = InteractiveCoFit(wa, flux, sig, contpoints,
-                                   co=co_init, fig=fig, anchor=anchor, numguesspix=numguesspix)
+                                   co=co_init, fig=fig, anchor=anchor, numguesspix=numguesspix, outfknots=outfknots)
 
         # wait until the interactive fitting has finished
         while not wrapper.finished: